chore(openai): migrate refactor changes
